more type corrections

+++ common/src/backend.rs
@@ -58,7 +58,7 @@ pub struct CreateTaskResponse {
     /// id of task
     pub task_id: TaskID,
     /// id of request
-    pub req_id: u64,
+    pub req_id: i32,
 }
 
 /// reqwest::post("/tasks").body(CreateTaskRequest {})
